refactor: factor out selenium from provider
I did this mainly for testing and for cohesion, it make no sense for
the provider which should represent a speedtest service to have a hard
dependency to selenium.

Now the a provider expect an interface which should be easier to mock or
use another framework like pyppeteer or others.

Now time to write some tests.
