+++ main.py
@@ -5,6 +5,7 @@ import argparse
 import sys
 
 from speed_sleuth.browser import BrowserFactory
+from speed_sleuth.driver.selenium import SeleniumDriver
 from speed_sleuth.provider.speedofme import *  # to import object into the global table.
 from speed_sleuth.provider.speedtest import *  # to import object into the global table.
 
@@ -19,11 +20,13 @@ args = parser.parse_args()
 
 browser = BrowserFactory.get_browser()
 
+driver = SeleniumDriver(browser)
+
 for provider in args.providers:
     class_name = provider.capitalize()
     ProviderClass = globals().get(class_name)
     if ProviderClass:
-        instance = ProviderClass(browser)
+        instance = ProviderClass(driver)
         for i in range(args.count):
             instance.run(f"{provider}-{i}.png")
     else:
